Gear Up: Integrated latest dev into stable.
[git-p4: depot-paths = "//depot/traktor/": change = 15010]

+++ code/Amalgam/Engine/Action/Classes/As_traktor_amalgam_DisplayMode.cpp
@@ -59,7 +59,7 @@ void As_traktor_amalgam_DisplayMode::DisplayMode_get_availableModes(flash::CallA
 {
 	RefArray< AsDisplayMode > displayMode = AsDisplayMode::getAvailableModes(m_environment);
 	
-	Ref< flash::Array > displayModeArr = new flash::Array(displayMode.size());
+	Ref< flash::Array > displayModeArr = new flash::Array(uint32_t(displayMode.size()));
 	for (uint32_t i = 0; i < displayMode.size(); ++i)
 	{
 		displayModeArr->push(flash::ActionValue(
