+++ code/Core/Misc/TString.cpp
@@ -14,7 +14,7 @@ std::wstring mbstows(const IEncoding& encoding, const std::string& s)
 	const char* cs = s.c_str();
 	for (uint32_t i = 0; i < s.length(); )
 	{
-		uint32_t nb = std::min< uint32_t >(IEncoding::MaxEncodingSize, s.length() - i);
+		uint32_t nb = std::min< uint32_t >(IEncoding::MaxEncodingSize, uint32_t(s.length() - i));
 		if ((r = encoding.translate((const uint8_t*)&cs[i], nb, ec)) < 0)
 			break;
