+++ code/Net/Replication/RelayPeers.cpp
@@ -94,7 +94,7 @@ bool RelayPeers::setPrimaryPeerHandle(handle_t handle)
 uint32_t RelayPeers::getPeers(std::vector< PeerInfo >& outPeers) const
 {
 	outPeers = m_info;
-	return outPeers.size();
+	return uint32_t(outPeers.size());
 }
 
 int32_t RelayPeers::receive(void* data, int32_t size, handle_t& outFromHandle)
