+++ code/Sound/Filters/GroupFilter.cpp
@@ -69,7 +69,7 @@ Ref< IFilterInstance > GroupFilter::createInstance() const
 void GroupFilter::apply(IFilterInstance* instance, SoundBlock& outBlock) const
 {
 	GroupFilterInstance* gfi = static_cast< GroupFilterInstance* >(instance);
-	uint32_t nfilters = m_filters.size();
+	uint32_t nfilters = uint32_t(m_filters.size());
 	for (uint32_t i = 0; i < nfilters; ++i)
 		m_filters[i]->apply(gfi->m_instances[i], outBlock);
 }
