+++ code/Sound/SoundDriverNull.cpp
@@ -22,7 +22,7 @@ void SoundDriverNull::destroy()
 
 void SoundDriverNull::wait()
 {
-	long ms = m_desc.frameSamples * 1000L / m_desc.sampleRate;
+	int32_t ms = int32_t(m_desc.frameSamples * 1000L / m_desc.sampleRate);
 	ThreadManager::getInstance().getCurrentThread()->sleep(ms);
 }
